requirements.txt
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
python-dotenv==1.0.1
requests==2.32.3
razorpay==1.4.2
reportlab==4.2.2
PyJWT==2.9.0
